Se mueve la logica para obtener los datos del usuario del login al contexto
